Add List display by drone

+++ Delete_Photo.php
@@ -1,17 +1,17 @@
-
 <?php
 include "config.php";
-    if (isset($_GET['d'])):
-        $stmt = $mysqli->prepare("DELETE content,content_dron_category FROM content 
-            INNER JOIN content_dron_category ON content.id_content = content_dron_category.id_content
-            WHERE content.id_content=?");
-        $stmt->bind_param('s', $id);  
+    if (isset($_GET['d']) && isset($_GET['f'])):
         $id = $_GET['d'];
-         
+        $file = $_GET['f'];
+
+        $query = "SELECT file from Images where id='$id'";
+        $filename = mysqli_fetch_array(mysqli_query($mysqli, $query))['file'];
+
+        $query2 = "DELETE FROM Images WHERE id='". $id ."'";
+        $stmt = $mysqli->prepare($query2);
+        $stmt->bind_param('s', $id);
         if ($stmt->execute()):
-            
             $mysqli->close();
-        $filename=$_GET['f'];
             unlink('./php/album/'.$filename);
             echo "<script>location.href='principal.php'</script>";
         else:
